Improved import structure
